requirements.txt
Flask==3.1.2
Flask_Cors==4.0.0
Flask_Limiter==3.5.0
joblib==1.3.2
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.1
PyJWT==2.8.0
PyJWT==2.11.0
python_bcrypt==0.3.2
python_whois==0.9.4
scikit_learn==1.4.0